[DEV] Welcome Member Banner Bug (#44)

+++ src/commands/misc/help.ts
@@ -6,14 +6,10 @@ import {
   RuleBuilder,
 } from '@typeit/discord';
 import { environment } from '../../utils/environment';
-import {
-  commandOverrides,
-  commandTypes,
-  commandHelpTypes,
-} from '../../data/help';
+import { commandOverrides, commandHelpTypes } from '../../data/help';
 import Utility from '../../utils/utility';
 import { EmbedFieldData, Message, MessageEmbed } from 'discord.js';
-import { CommandType } from '../../types/help';
+import { CommandItem, CommandType } from '../../types/help';
 
 export class Help {
   /**
@@ -23,9 +19,9 @@ export class Help {
    */
   private createBaseMessage(
     command: CommandMessage,
-    type?: CommandType
+    type?: CommandItem
   ): MessageEmbed {
-    const t = type ? ` ${Utility.captaliseFirstLetter(type)}` : '';
+    const t = type ? ` ${type.name}` : '';
     return new MessageEmbed()
       .setColor(10181046)
       .setAuthor(
@@ -41,7 +37,7 @@ export class Help {
   private async createHelpStatus(
     command: CommandMessage,
     allCommands: CommandInfos<any, RuleBuilder>[],
-    type?: CommandType
+    type?: CommandItem
   ): Promise<Message | void> {
     if (type) {
       const fields = allCommands.map((c) => {
@@ -71,10 +67,10 @@ export class Help {
           value: `\`${fullCommand}\``,
         };
       });
-      const restrictedCommand = commandHelpTypes.filter((c) => c.restrict);
+
       if (!Utility.isAdmin(command)) {
-        fields = fields.filter((c) =>
-          restrictedCommand.find((r) => r.name === c.name)
+        fields = fields.filter(
+          (c) => !commandHelpTypes.find((r) => r.name === c.name)?.restrict
         );
       }
 
@@ -88,12 +84,12 @@ export class Help {
    * Fetch Commands
    * @param type
    */
-  fetchCommands(type?: CommandType): CommandInfos<any, RuleBuilder>[] {
+  fetchCommands(type?: CommandItem): CommandInfos<any, RuleBuilder>[] {
     const commands = Client.getCommands();
 
     if (type) {
       const commandNames = commandOverrides
-        .filter((c) => c.type === type)
+        .filter((c) => c.type === type.name.toLowerCase())
         .map((c) => c.name);
       return commands.filter(
         (c) => commandNames.indexOf(c.commandName as string) > -1
@@ -111,13 +107,14 @@ export class Help {
    */
   @Command('help')
   async helpInit(command: CommandMessage): Promise<Message | void> {
-    const type = Utility.getOptionFromCommand(
-      command.content,
-      2,
-      ' '
-    ) as CommandType;
+    const type = (
+      Utility.getOptionFromCommand(command.content, 2, ' ') as CommandType
+    ).toLowerCase();
 
-    if (type && !commandTypes.find((c) => c === type)) {
+    const commandGroup = commandHelpTypes.find(
+      (c) => c.name.toLowerCase() === type
+    );
+    if (type && !commandGroup) {
       await command.delete();
       return command.channel
         .send(
@@ -126,16 +123,18 @@ export class Help {
         .then((m) => m.delete({ timeout: 5000 }));
     }
 
-    if (!Utility.isAdmin(command)) {
+    if (!Utility.isAdmin(command) && commandGroup?.restrict) {
       await command.delete();
       return command.channel
         .send(`**You don't have the permissions for this**`)
         .then((m) => m.delete({ timeout: 5000 }));
     }
 
-    const allCommands = this.fetchCommands(type);
-    return this.createHelpStatus(command, allCommands, type).catch(() => {
-      command.reply(environment.error);
-    });
+    const allCommands = this.fetchCommands(commandGroup);
+    return this.createHelpStatus(command, allCommands, commandGroup).catch(
+      () => {
+        command.reply(environment.error);
+      }
+    );
   }
 }
